+++ geg/src/window/window.hpp
@@ -21,8 +21,8 @@ namespace Geg {
 		void onUpdate();
 		[[nodiscard]] GLFWwindow* getRawWindow() const { return windowPtr; };
 		[[nodiscard]] GraphicsContext* getGraphicsContext() const { return context; };
-		[[nodiscard]] const int getWidth() const { return info.width; };
-		[[nodiscard]] const int getHeight() const { return info.height; };
+		[[nodiscard]] const float getWidth() const { return info.width; };
+		[[nodiscard]] const float getHeight() const { return info.height; };
 		[[nodiscard]] const bool getVSync() const { return info.VSync; }
 		void setVsync(bool state);
 		void setEventCallback(const std::function<void(Event&)>& cb) { info.eventCallback = cb; }
@@ -31,7 +31,7 @@ namespace Geg {
 		GLFWwindow* windowPtr;
 		GraphicsContext* context;
 		struct WindowInfo {
-			int width, height;
+			float width, height;
 			bool VSync = true;
 			std::string name;
 			std::function<void(Event&)> eventCallback;
